+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [ADDED] - `ValidationError.API_CODE` static writable property (which defaults to `validationFailure`) - sets `api_code` value for validation error http 400 response
 * [ADDED] - Ajv option `allErrors=true` is now supported. if enabled, validation response will include a collection of failures
+* [ADDED] - `bi-service test:config` command
 
 ## 1.4.0